Inflation Comparision between Turkey and EU  / OECD Countries
I got the data from https://data.oecd.org/price/inflation-cpi.htm. What I observe in this data set is that Turkey faces traumatic and higher inflation than OECD and EU countries. In the last 8 months, The Central Bank of Turkey lowered its primary interest rate, which triggered a jump in the exchange rates. As a result, Turkey became one of the main inflationist countries in the world.
I got the data from https://data.oecd.org/price/inflation-cpi.htm. What I observe in this data set is that Turkey faces traumatic and higher inflation than OECD and EU countries. In the last 8 months, The Central Bank of Turkey lowered its primary interest rate, which triggered a jump in the exchange rates. As a result, Turkey became one of the main inflationist countries in the world.